@@ -1,7 +1,8 @@
 package wendu.jsbdemo;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import wendu.dsbridge.DWebView;
 
