@@ -2,28 +2,18 @@ package wendu.jsbdemo;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 
-import com.tencent.smtt.sdk.QbSdk;
+import androidx.appcompat.app.AppCompatActivity;
+
+
 
 public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        QbSdk.initX5Environment(this, new QbSdk.PreInitCallback() {
-            @Override
-            public void onCoreInitFinished() {
-                Log.d("jsbridge","onCoreInitFinished");
-            }
-
-            @Override
-            public void onViewInitFinished(boolean b) {
-                Log.d("jsbridge","onCoreInitFinished");
-            }
-        });
         setContentView(R.layout.activity_main);
         findViewById(R.id.callJs).setOnClickListener(new View.OnClickListener() {
             @Override
@@ -38,6 +28,12 @@ public class MainActivity extends AppCompatActivity {
             }
         });
         findViewById(R.id.fly).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                startActivity(new Intent(MainActivity.this,WrokWithFlyioTestActivity.class));
+            }
+        });
+        findViewById(R.id.testx5).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 startActivity(new Intent(MainActivity.this,WrokWithFlyioTestActivity.class));
