@@ -1,8 +1,9 @@
 package wendu.jsbdemo;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.webkit.JavascriptInterface;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import org.json.JSONObject;
 
